Renamed mac_changer file to wireless
